+++ src/format/pg/parser.js
@@ -206,7 +206,7 @@ function peg$parse(input, options) {
   var peg$r3 = /^[^ \t\n\r<>"{}|\^`\\]/;
   var peg$r4 = /^[:([#\-]/;
   var peg$r5 = /^[^`]/;
-  var peg$r6 = /^[^\0-\x1F"'\\]/;
+  var peg$r6 = /^[^\0-\b\v\f\x0E-\x1F"'\\]/;
   var peg$r7 = /^["'\/\\]/;
   var peg$r8 = /^[1-9]/;
   var peg$r9 = /^[0-9]/;
@@ -238,7 +238,7 @@ function peg$parse(input, options) {
   var peg$e21 = peg$literalExpectation("'", false);
   var peg$e22 = peg$literalExpectation("`", false);
   var peg$e23 = peg$classExpectation(["`"], true, false);
-  var peg$e24 = peg$classExpectation([["\0", "\x1F"], "\"", "'", "\\"], true, false);
+  var peg$e24 = peg$classExpectation([["\0", "\b"], "\v", "\f", ["\x0E", "\x1F"], "\"", "'", "\\"], true, false);
   var peg$e25 = peg$literalExpectation("\\", false);
   var peg$e26 = peg$classExpectation(["\"", "'", "/", "\\"], false, false);
   var peg$e27 = peg$literalExpectation("b", false);